refactor(ffi): migrate from hand-written C bindings to type-safe CXX bridge
Replace game_bindings.cpp — 373 lines of extern "C" with raw opaque
pointers and manual new/delete — with a type-safe CXX bridge. Rust now
holds a UniquePtr<GameBridge> with shared types visible to both languages,
no unsafe pointer handling, and no manual lifetime management.

Event drain is redesigned around ownership transfer: a single ECS lookup
moves the entire event vector out via std::exchange into a
UniquePtr<EventQueue>, replacing a per-event registry scan that required
2N FFI calls per tick.

Vec3 ↔ Vector3D conversion is zero-cost via reinterpret_cast, guarded by
static_assert layout checks at compile time.

build.rs is rewritten around cxx_build with conditional debug/release
flags, cross-language thin LTO, -march=x86-64-v3, and directory-level
rebuild triggers. Release profile adds codegen-units=1 and symbol stripping.

Also: C++ warning cleanup across game-core, deque → vector for the event
component, and a `make bear` target for IDE compile_commands.json generation.

+++ game-core/src/systems/CombatSystem.hpp
@@ -59,7 +59,7 @@ namespace ArenaGame {
 // stageMultiplier comes from AttackStage::damageMultiplier or SkillDefinition::dmgMultiplier.
 inline float calculateCombatDamage(const Components::CombatController& cc, float stageMultiplier) {
 	float dmg = cc.baseDamage * stageMultiplier * cc.damageMultiplier;
-	bool isCrit = (static_cast<float>(std::rand()) / RAND_MAX) < cc.criticalChance;
+	bool isCrit = (static_cast<float>(std::rand()) / static_cast<float>(RAND_MAX)) < cc.criticalChance;
 	return isCrit ? dmg * cc.criticalMultiplier : dmg;
 }
